Added unit tests for game
