+++ data/games_authors.json
@@ -1,1 +1,1 @@
-[{"publish_date":"12/10/2019","last_played_at":"12/09/2020","multiplayer":true,"author":{"first_name":"Tazan","last_name":"Kavuma"}},{"publish_date":"12/01/2020","last_played_at":"29/02/2021","multiplayer":false,"author":{"first_name":"Vincent","last_name":"Ondiff"}}]
+[{"publish_date":"12/10/2019","last_played_at":"12/09/2020","multiplayer":true,"author":{"first_name":"Tazan","last_name":"Kavuma"}},{"publish_date":"12/01/2020","last_played_at":"29/02/2021","multiplayer":false,"author":{"first_name":"Vincent","last_name":"Ondiff"}},{"publish_date":"12/01/1990","last_played_at":"10/09/2020","multiplayer":true,"author":{"first_name":"Tazan","last_name":"Kavuma"}}]
